requirements/test.txt
flake8>=2.4,<3
mock>=3.0,<4
pytest>=2.7,<3
requests-mock>=0.6.0
analyzere>=0.5.2,<1
